Addresses sources of UB in generator
- Enables lots of warnings with -Wall -Wextra -Wpedantic
- Makes warning into errors with -Werror
  - no-strict-aliasing for now
- Enables undefined behavior sanitizer

This fixes the existing warnings

[ROCm/mxDataGenerator commit: 9be95799266dc55f9eaa4a71a84ca2327d432506]

+++ shared/mxdatagenerator/test/f32_test.cpp
@@ -735,8 +735,8 @@ TEST_F(f32_test, getDataMaxSubnorm)
 {
 
     float subMax = 0;
-    for(int i = 0; i < f32::dataInfo.mantissaBits; i++)
-        subMax += std::pow(2, -(i + 1));
+    for(size_t i = 0; i < f32::dataInfo.mantissaBits; i++)
+        subMax += std::pow(2, -int32_t(i + 1));
 
     subMax *= std::pow(2, -126);
