+++ shared/mxdatagenerator/test/ocp_e5m2_mxfp8_test.cpp
@@ -199,7 +199,7 @@ protected:
     {
         double closestDiff = UINT32_MAX;
 
-        for(int i = 0; i < e5m2ValuesOCP.size(); i++)
+        for(size_t i = 0; i < e5m2ValuesOCP.size(); i++)
         {
             if(std::isnan(e5m2ValuesOCP[i]))
                 continue;
@@ -761,7 +761,7 @@ TEST_F(ocp_e5m2_mxfp8_test, isSubnorm)
 {
     uint8_t temp[] = {DGen::Constants::E8M0_1, 0b0};
 
-    for(int i = 0; i < e5m2ValuesOCP.size(); i++)
+    for(size_t i = 0; i < e5m2ValuesOCP.size(); i++)
     {
         uint8_t data = static_cast<uint8_t>(i) & 0xff;
         temp[1]      = data;
@@ -824,10 +824,10 @@ TEST_F(ocp_e5m2_mxfp8_test, isGreater)
 
 TEST_F(ocp_e5m2_mxfp8_test, toFloatAllScalesAllValues)
 {
-    for(int i = 0; i < e8m0Values.size(); i++)
+    for(size_t i = 0; i < e8m0Values.size(); i++)
     {
         float ref = e8m0Values[i];
-        for(int j = 0; j < e5m2ValuesOCP.size(); j++)
+        for(size_t j = 0; j < e5m2ValuesOCP.size(); j++)
         {
             float  res      = toFloat<DT>(e8m0Bits, e5m2BitsOCP, i, j);
             double expected = ref * e5m2ValuesOCP[j];
@@ -849,11 +849,11 @@ TEST_F(ocp_e5m2_mxfp8_test, toFloatAllScalesAllValues)
 
 TEST_F(ocp_e5m2_mxfp8_test, toDoubleAllScalesAllValues)
 {
-    for(int i = 0; i < e8m0Values.size(); i++)
+    for(size_t i = 0; i < e8m0Values.size(); i++)
     {
         double ref = e8m0Values[i];
 
-        for(int j = 0; j < e5m2ValuesOCP.size(); j++)
+        for(size_t j = 0; j < e5m2ValuesOCP.size(); j++)
         {
             double res      = toDouble<DT>(e8m0Bits, e5m2BitsOCP, i, j);
             double expected = ref * e5m2ValuesOCP[j];
